+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ New features
 * Automatically add a dummy DV column in NONMEM if needed. This will make it possible to do simulations in NONMEM without having a DV in the dataset.
 * Add :code:`modeling.export_model_files`
 * Add :code:`modeling.map_eta_parameters` for creation of mappings between connected individual parameters, etas or omegas.
+* Add :code`modeling.infer_datatypes` to infer and check if data columns can be converted a simpler datatype (e.g. int32)
 
 Changes
 =======